Added new  stuff
Added scroll  view
add button for the addition of the items to the cart,
then added a new gradient fill,
then added a new alert functions for any overloads in function.

Made a function that made it so that when the user reaches over six courses they will not be be able to enter more  courses,
Made styles for the info bar and  the linear gradient.
Deleted the linking of the courses in the page in the First Aid course page.